requirements.txt
fastapi==0.112.2
llama-index-llms-groq==0.2.0
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
typesense==0.20.0
uvicorn==0.30.6
google-api-python-client==2.143.0